fix: Fix base branch identifier for pull requests in release workflow
When releasing from branches containing a `/` the gh CLI call to open a
pull request specified an invalid base branch by passing `--base
"${GITHUB_REF##*/}"` which would chomp everything before a `/` in the
current GITHUB_REF (which contains refs/heads/<branchname> or
refs/tags/<tagname>). This would mangle the branch name if it contains a
`/` character so we use GITHUB_REF_NAME instead which should not contain
any prefix.

Signed-off-by: Paolo Chila <paolo.chila@dynatrace.com>